requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
httpx>=0.27.0
mcp>=1.0.0
python-dotenv>=1.0.0
